backend/requirements.txt
# Web Framework
fastapi>=0.104.0,<1.0
uvicorn[standard]>=0.24.0,<1.0
python-multipart>=0.0.6,<1.0

# Database
sqlalchemy>=2.0.23,<3.0
psycopg2-binary>=2.9.9,<3.0
asyncpg>=0.29.0,<1.0

# Document Parsing
PyMuPDF>=1.23.0,<2.0
python-docx>=1.1.0,<2.0
Pillow>=10.1.0,<13.0

# Security
python-jose[cryptography]>=3.3.0,<4.0
passlib[bcrypt]>=1.7.4,<2.0

# Async SQLite (local dev)
aiosqlite>=0.19.0,<1.0

# Report Export
openpyxl>=3.1.2,<4.0
reportlab>=4.0.0,<5.0

# Utils
pydantic>=2.5.0,<3.0
pydantic-settings>=2.1.0,<3.0
